+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### New Features
 
+- Added a `ReusableAborter`
+- Added `metadata` field for `AbortError` and `TimeoutError`.
 - Added utility functions:
   - catchAbortError
   - isAbortSignal